ID:103 Issue: code optimization and Activate code for user. Fixed: Fixed the issue. Test: Tested using insomnia.
