+++ Server/app.ts
@@ -3,7 +3,7 @@ import cors from "cors"
 import cookieParser from "cookie-parser";
 import {ErrorMiddleware} from "./middleware/error"
 require("dotenv").config();
-import userRouter from "./routes/user.route"
+import userRouter from "./routes/userRoute"
 export const app = express();
 
 //body parser
@@ -14,28 +14,12 @@ app.use(cookieParser())
 app.use(cors({origin:process.env.ORIGIN}))
 
 app.use("/user",userRouter);
-//test api
-app.get("/hello",(req:Request,res:Response,next:NextFunction)=>{
-    res.json({
-        status:500,
-        message:"Working"
-    })
-})
 
-app.all("*/:id",(req:Request,res:Response,next:NextFunction)=>{
-    console.log("id",req.params.id);
-    
+app.all("*",(req:Request,res:Response,next:NextFunction)=>{
     const err = new Error(`Route ${req.originalUrl} not found`) as any;
     err.statusCode = 400;      
     console.log("er:",err)
     next(err)
 })
-
-app.use((req, res, next) => {
-    console.log("Check bala");
-     
-
-
-})
-
+console.log("hello");
 app.use(ErrorMiddleware)
