--- Server/controllers/user.controller.ts
@@ -1,77 +1,0 @@
-require("dotenv").config();
-import { Request, Response, NextFunction } from "express";
-import userModel from "../models/user.model";
-import Jwt, { Secret } from "jsonwebtoken";
-import ErrorHandler from "../utils/ErrorHandler";
-import { CatchAsyncError } from "../middleware/CatchAsyncErrors";
-import ejs, { renderFile } from "ejs";
-import path from "path";
-import sendMail from "../utils/sendMail";
-
-//create new user
-interface ICreationBody {
-  name: string;
-  email: string;
-  password: string;
-  avatar?: string;
-}
-
-export const CreationUser = CatchAsyncError(
-  async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const { name, email, password } = req.body;
-      const isEmailExist = await userModel.findOne({ email });
-      if (isEmailExist) {
-        next(new ErrorHandler("Email already exist", 400));
-      }
-      const user: ICreationBody = { name, email, password };
-      const activationToken = createActivationToken(user);
-      const activationCode = activationToken.activationCode;
-      const data = { user: { name: user.name }, activationCode };
-      const html = await ejs.renderFile(
-        path.join(__dirname, "../mails/activation-mail.ejs"),
-        data
-      );
-
-      try {
-        await sendMail({
-          email: user.email,
-          subject: "Activate your account",
-          template: "activation-mail.ejs",
-          data,
-        });
-        res.status(201).json({
-          success: true,
-          message: `Please check your email ${user.email} to activate your account`,
-          activationToken: activationToken.token,
-        });
-      } catch (error: any) {
-        return next(new ErrorHandler(error.message, 400));
-      }
-    } catch (error: any) {
-      return next(new ErrorHandler(error.message, 400));
-    }
-  }
-);
-
-interface IActivationToken {
-  token: string;
-  activationCode: string;
-}
-
-export const createActivationToken = (user: any): IActivationToken => {
-  const activationCode = Math.floor(1000 + Math.random() * 9000).toString();
-  const token = Jwt.sign(
-    {
-      user,
-      activationCode,
-    },
-    process.env.ACTIVATION_SECRET as Secret,
-    {
-      expiresIn: "5m",
-    }
-  );
-  return { token, activationCode };
-};
-
-export default CreationUser;
